Move UserCoursesActions and UserCoursesOptions components to outside folder
