+++ src/components/UserCourses/UserCoursesActions.jsx
@@ -1,8 +1,8 @@
 import { faAngleDown } from "@fortawesome/free-solid-svg-icons";
 import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
 import React, { useState } from "react";
-import SortUserCourses from "./SortUserCourses";
-import AllFilters from "../../../../shared/Filters/AllFilters";
+import SortUserCourses from "../../shared/Sort";
+import AllFilters from "../../shared/Filters/AllFilters";
 
 const UserCoursesActions = React.memo(() => {
   const [areActionsOpen, setAreActionsOpen] = useState(false);
